+++ be/src/data.csv
@@ -1,0 +1,4 @@
+name,age,city
+John Doe,30,New York
+Jane Smith,25,San Francisco
+Bob Johnson,40,Los Angeles
