+++ be/src/modules/idea/idea.service.ts
@@ -8,9 +8,7 @@ import { CommentService } from '@modules/comment/comment.service';
 import { IdeaFileService } from '@modules/idea-file/idea-file.service';
 import { ReactionService } from '@modules/reaction/reaction.service';
 import { EventService } from '@modules/event/event.service';
-import { HttpException, HttpStatus, Injectable, StreamableFile } from '@nestjs/common';
-import type { Response, Request } from 'express';
-import * as send from 'send';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { EIsDelete } from 'enum';
 import { EUserRole } from 'enum/default.enum';
@@ -20,10 +18,7 @@ import { VAddComment } from 'global/dto/addComment.dto';
 import { VCreateIdeaDto } from 'global/dto/create-idea.dto';
 import { VCreateReactionDto } from 'global/dto/reaction.dto';
 import { VUpdateIdeaDto } from 'global/dto/update-idea.dto';
-import { join } from 'path';
-import * as fs from 'fs';
 import * as moment from 'moment';
-import {stringify} from 'csv-stringify';
 import { Idea } from 'src/core/database/mysql/entity/idea.entity';
 import {
   Repository,
@@ -515,86 +510,6 @@ export class IdeaService {
     }
 
     return this.commentService.getIdeaCommentsByParent(idea_id, parent_id);
-  }
-
-  downloadIdeasByEvent(
-    userData: IUserData,
-    event_id: number,
-    res: Response,
-    req: Request,
-    entityManager?: EntityManager,
-  ) {
-    const ideaRepository = entityManager
-      ? entityManager.getRepository<Idea>('idea')
-      : this.ideaRepository;
-
-    if (userData.role_id != EUserRole.QA_MANAGER) {
-      throw new HttpException(
-        ErrorMessage.DATA_DOWNLOAD_PERMISSION,
-        HttpStatus.BAD_REQUEST,
-      );
-    }
-    const temp = join(process.cwd(), 'package.json');
-    send(req, temp).pipe(res);
-    return;
-    // const temp = fs.createReadStream(join(process.cwd(), 'package.json'));
-    // return new StreamableFile(temp);
-
-    // const event = await this.eventService.getEventById(event_id);
-    
-    // if(!event) {
-    //   throw new HttpException(
-    //     ErrorMessage.event_NOT_EXIST,
-    //     HttpStatus.BAD_REQUEST,
-    //   );
-    // }
-
-    // if(event.final_closure_date) {
-    //   throw new HttpException(
-    //     ErrorMessage.DATA_DOWNLOAD_DATE_TIME,
-    //     HttpStatus.BAD_REQUEST,
-    //   );
-    // }
-    
-    const data = [
-      [ 'John Doe', 30, 'New York' ],
-      [ 'Jane Smith', 25, 'San Francisco' ],
-      [ 'Bob Johnson', 40, 'Los Angeles' ],
-    ];
-    
-    const fileName = "data.csv";
-    const path = join(process.cwd(), 'src', fileName);
-    
-    const writableStream = fs.createWriteStream(path);
-    const columns = [
-      "name",
-      "age",
-      "city",
-    ];
-    
-    // try {
-      const stringifier = stringify({ header: true, columns: columns });
-      data.forEach(d => {
-        stringifier.write(d);
-      });
-
-      // stringifier.pipe(res);
-
-      res.set({
-        'Content-Type': 'text/csv',
-        'Content-Disposition': 'attachment; filename="data.csv"',
-      })
-      // res.sendFile(path);
-      const file = fs.createReadStream(path);
-      // file.pipe(res);
-      res.send(file.pipe(res));
-      // return new StreamableFile(file);
-    // } catch (error) {
-    //   throw new HttpException(
-    //     ErrorMessage.DATA_DOWNLOAD_FAILED,
-    //     HttpStatus.BAD_REQUEST,
-    //   );
-    // }
   }
 
   async createComment(userData: IUserData, idea_id: number, body: VAddComment) {
